\name{microTRI}
\alias{microTRI}

\title{microTRI}

\description{This package is a R shiny application dedicated to small vertebrates analyses.}

\usage{
microTRI()
}

\details{This is a launcher function for the `microTRI` application.}

\value{Launch the `microTRI` Shiny application.}

\author{Aurelien Royer <aurelien.royer at u-bourgogne.fr>}

\seealso{
	\code{shiny}
	\code{plotly}
	\code{ggplot2}
}

\references{
Royer, A. - github page
}

\examples{
	if(interactive()){microTRI()}
}
